+++ src/IV-2_6/button2_6.c
@@ -357,7 +357,7 @@ TextButton::~TextButton() {
 }
 
 PushButton::PushButton(
-    const char* str, ButtonState* s, int v
+    const char* str, ButtonState* s, long v
 ) : TextButton(str, s, (void*)v) {
     Init();
 }
@@ -369,7 +369,7 @@ PushButton::PushButton(
 }
 
 PushButton::PushButton(
-    const char* name, const char* str, ButtonState* s, int v
+    const char* name, const char* str, ButtonState* s, long v
 ) : TextButton(name, str, s, (void*)v) {
     Init();
 }
@@ -445,7 +445,7 @@ static Bitmap* radioChosen;
 static Bitmap* radioBoth;
 
 RadioButton::RadioButton(
-    const char* str, ButtonState* s, int v
+    const char* str, ButtonState* s, long v
 ) : TextButton(str, s, (void*)v) {
     Init();
 }
@@ -457,7 +457,7 @@ RadioButton::RadioButton(
 }
 
 RadioButton::RadioButton(
-    const char* name, const char* str, ButtonState* s, int v
+    const char* name, const char* str, ButtonState* s, long v
 ) : TextButton(name, str, s, (void*)v) {
     Init();
 }
@@ -533,7 +533,7 @@ void RadioButton::Refresh() {
 }
 
 CheckBox::CheckBox(
-    const char* str, ButtonState* s, int on, int off
+    const char* str, ButtonState* s, long on, long off
 ) : TextButton(str, s, (void*)on) {
     Init((void*)off);
 }
@@ -545,7 +545,7 @@ CheckBox::CheckBox(
 }
 
 CheckBox::CheckBox(
-    const char* name, const char* str, ButtonState* s, int on, int off
+    const char* name, const char* str, ButtonState* s, long on, long off
 ) : TextButton(name, str, s, (void*)on) {
     Init((void*)off);
 }
